refactoring: remove QuizResultsViewModel, move more logic to Presenter
