--- MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,14 +1,0 @@
-//
-//  QuizResultsViewModel.swift
-//  MovieQuiz
-//
-//  Created by Roman Romanov on 31.03.2024.
-//
-
-import Foundation
-
-struct QuizResultsViewModel {
-  let title: String
-  let text: String
-  let buttonText: String
-}
