+++ MovieQuiz/Presentation/MovieQuizPresenter.swift
@@ -27,19 +27,6 @@ final class MovieQuizPresenter {
     
     // MARK: - public functions
     
-    func loadData() {
-        viewController?.showLoadingIndicator()
-        Task {
-            await questionFactory?.loadData()
-        }
-    }
-    
-    func restartGame() {
-        correctAnswers = 0
-        currentQuestionIndex = 0
-        questionFactory?.requestNextQuestion()
-    }
-    
     func showAnswerResult(isCorrect: Bool) {
         guard let currentQuestion else { return }
         
@@ -60,20 +47,6 @@ final class MovieQuizPresenter {
         }
     }
     
-    func makeResultsMessage() -> String {
-        statisticService.store(correct: correctAnswers, total: questionsAmount)
-        
-        let bestGame = statisticService.bestGame
-        let message = """
-            Ваш результат: \(correctAnswers)/\(questionsAmount)
-            Количество сыгранных квизов: \(statisticService.gamesCount)
-            Рекорд \(bestGame.correct)/\(bestGame.total) \(bestGame.date.dateTimeString)
-            Срендяя точность \(String(format: "%.2f", statisticService.totalAccuracy))%
-        """
-        
-        return message
-    }
-    
     func convert(model: QuizQuestion) -> QuizStepViewModel {
         QuizStepViewModel(
             image: UIImage(data: model.image) ?? UIImage(),
@@ -85,6 +58,19 @@ final class MovieQuizPresenter {
 // MARK: - private functions
 
 extension MovieQuizPresenter {    
+    private func loadData() {
+        viewController?.showLoadingIndicator()
+        Task {
+            await questionFactory?.loadData()
+        }
+    }
+    
+    private func restartGame() {
+        correctAnswers = 0
+        currentQuestionIndex = 0
+        questionFactory?.requestNextQuestion()
+    }
+    
     private func isLastQuestion() -> Bool {
         currentQuestionIndex == questionsAmount - 1
     }
@@ -107,7 +93,31 @@ extension MovieQuizPresenter {
             return
         }
         
-        viewController?.showQuizResults()
+        let message = makeResultsMessage()
+        let model = AlertModel(
+            title: "Этот раунд окончен!",
+            message: message,
+            buttonText: "Сыграть ещё раз") { [weak self] in
+                guard let self else { return }
+                
+                self.restartGame()
+            }
+        
+        viewController?.showAlert(model: model)
+    }
+    
+    private func makeResultsMessage() -> String {
+        statisticService.store(correct: correctAnswers, total: questionsAmount)
+        
+        let bestGame = statisticService.bestGame
+        let message = """
+            Ваш результат: \(correctAnswers)/\(questionsAmount)
+            Количество сыгранных квизов: \(statisticService.gamesCount)
+            Рекорд \(bestGame.correct)/\(bestGame.total) \(bestGame.date.dateTimeString)
+            Срендяя точность \(String(format: "%.2f", statisticService.totalAccuracy))%
+        """
+        
+        return message
     }
 }
 
@@ -132,6 +142,20 @@ extension MovieQuizPresenter: QuestionFactoryDelegate {
     
     func didFailToLoadData(with error: Error) {
         let message = error.localizedDescription
-        viewController?.showNetworkError(message: message)
+        
+        viewController?.hideLoadingIndicator()
+        
+        let model = AlertModel(
+            title: "Ошибка",
+            message: message,
+            buttonText: "Попробовать ещё раз") { [weak self] in
+                guard let self else { return }
+                
+                self.restartGame()
+                // load data one more time
+                self.loadData()
+            }
+        
+        viewController?.showAlert(model: model)
     }
 }
